feat(board): add coordinate conversion and piece name management
- Introduced `board-coords.ts` for algebraic to internal coordinate conversion and vice versa, enhancing move parsing efficiency.
- Added `piece-names.ts` to manage canonical starting-position piece names for board replay and heatmap generation.
- Refactored utility functions in `utils.ts` to utilize the new coordinate and piece name modules, improving code organization and clarity.
- Updated `san-applier.ts` and `san-parser.ts` to directly use the new coordinate conversion functions, streamlining SAN processing.
- Implemented heatmap generation utilities in `heatmap-utils.ts` to facilitate analysis and visualization of chess data.
