+++ src/interfaces/index.ts
@@ -1,151 +1,33 @@
-import type { PlayerColor } from '../types';
+/**
+ * Public type surface for chessalyzer.js.
+ * Domain-specific definitions live in sibling modules; this file re-exports all of them
+ * so existing `from '../interfaces'` imports continue to work unchanged.
+ */
 
-export interface GameProcessorConfig {
-    hasFilter: boolean;
-    filter: (game: object) => boolean;
-    cntGames: number;
-}
+export type { MoveAction, CaptureAction, PromoteAction, Action } from './actions';
 
-export interface GameProcessorAnalysisConfig {
-    trackers: { move: Tracker[]; game: Tracker[] };
-    processedMoves: number;
-    processedGames: number;
-}
+export type {
+    Game,
+    Move,
+    ChessPiece,
+    SquareData,
+} from './game';
 
-export interface GameProcessorAnalysisConfigFull extends GameProcessorAnalysisConfig {
-    config: GameProcessorConfig;
-    trackerData: { name: string; cfg: TrackerConfig; path: string }[];
-    cntReadGames: number;
-    isDone: boolean;
-}
+export type {
+    Tracker,
+    TrackerConfig,
+    HeatmapData,
+    HeatmapAnalysisFunc,
+} from './tracker';
 
-export interface Game {
-    Result?: string;
-    ECO?: string;
-    moves: string[];
-}
+export type {
+    AnalysisConfig,
+    MultithreadConfig,
+    GameProcessorConfig,
+    GameProcessorAnalysisConfig,
+    GameProcessorAnalysisConfigFull,
+    GameAndMoveCount,
+    GameAndMoveCountFull,
+} from './analysis';
 
-export interface Move {
-    from: number[];
-    to: number[];
-}
-
-interface BaseAction {
-    type: 'move' | 'capture' | 'promote';
-    san: string;
-    player: PlayerColor;
-}
-
-export interface MoveAction extends BaseAction {
-    type: 'move';
-    piece: string;
-    from: number[];
-    to: number[];
-}
-
-export interface CaptureAction extends BaseAction {
-    type: 'capture';
-    takingPiece: string;
-    takenPiece: string;
-    on: number[];
-}
-
-export interface PromoteAction extends BaseAction {
-    type: 'promote';
-    to: string;
-    on: number[];
-}
-
-export type Action = MoveAction | CaptureAction | PromoteAction;
-
-export interface SquareData {
-    alg: string;
-    coords: number[];
-    piece: { color: PlayerColor; name: string };
-}
-
-export interface ChessPiece {
-    name: string; // not Piece since promoted pawns can be included here
-    color: PlayerColor;
-}
-
-export interface TrackerConfig {
-    profilingActive: boolean;
-}
-
-export interface MultithreadConfig {
-    batchSize: number;
-}
-
-export interface AnalysisConfig {
-    trackers?: Tracker[];
-    config?: {
-        cntGames?: number;
-        filter?: (data: unknown) => boolean;
-    };
-}
-
-export interface HeatmapData {
-    map: number[][];
-    min: number;
-    max: number;
-}
-
-export interface HeatmapAnalysisFunc {
-    (data: unknown, loopSqrData: SquareData, sqrData?: SquareData, optData?: unknown): number;
-}
-
-export interface Tracker {
-    type: string;
-    cfg: TrackerConfig;
-    time: number;
-    t0: number;
-    path?: string;
-    analyze: (arg: Game | Action[]) => void;
-    generateHeatmap: (
-        fun: string | HeatmapAnalysisFunc,
-        square?: string | number[],
-        optData?: unknown,
-    ) => HeatmapData;
-    generateComparisonHeatmap: (
-        compData: this,
-        fun: string | HeatmapAnalysisFunc,
-        square?: string | number[],
-        optData?: unknown,
-    ) => HeatmapData;
-    track: (arg: Game | Action[]) => void;
-    nextGame?: () => void;
-    finish?: () => void;
-    add?: (arg: this) => void;
-    /** Clear per-batch state when a worker reuses tracker instances. */
-    resetWorkerBatch?: () => void;
-}
-
-export interface GameAndMoveCount {
-    cntGames: number;
-    cntMoves: number;
-}
-
-export interface GameAndMoveCountFull extends GameAndMoveCount {
-    mps: number;
-}
-
-export interface WorkerInitData {
-    configs: {
-        trackerData: { name: string; cfg: TrackerConfig; path?: string }[];
-    }[];
-}
-
-/** Per-batch payload sent main → worker (tracker config lives in workerData). */
-export interface WorkerTaskData {
-    games: Game[];
-    idxConfig: number;
-}
-
-export interface WorkerMessage {
-    cntMoves: number;
-    cntGames: number;
-    idxConfig: number;
-    gameTrackers?: Tracker[];
-    moveTrackers?: Tracker[];
-}
+export type { WorkerInitData, WorkerTaskData, WorkerMessage } from './worker';
