+++ src/parsing/san-applier.ts
@@ -1,5 +1,5 @@
 import type { PieceToken } from '../types';
-import Utils from '../core/utils';
+import { algebraicToCoordsAt } from '../board/board-coords';
 import type SanContext from './san-context';
 
 /**
@@ -38,7 +38,7 @@ export default class SanApplier {
             end -= 2;
         }
 
-        const to = Utils.algebraicToCoordsAt(san, end);
+        const to = algebraicToCoordsAt(san, end);
         const from = this.ctx.fromBuf;
 
         if (san.charCodeAt(1) === 120) {
@@ -83,7 +83,7 @@ export default class SanApplier {
         const token = san.charAt(0) as PieceToken;
 
         const end = san.length;
-        const to = Utils.algebraicToCoordsAt(san, end);
+        const to = algebraicToCoordsAt(san, end);
 
         let restEnd = end - 2;
         let capture = false;
@@ -95,7 +95,7 @@ export default class SanApplier {
 
         let from: number[];
         if (restLen === 2) {
-            from = Utils.algebraicToCoordsAt(san, restEnd);
+            from = algebraicToCoordsAt(san, restEnd);
         } else if (restLen === 1) {
             const c = san.charCodeAt(1);
             const mustBeInCol = c >= 97 && c <= 104 ? c - 97 : null;
